+++ src/ts/VendingMachineCoinManager.ts
@@ -1,25 +1,10 @@
+import { CoinManager, coins } from './types/vendingMachineCoinManager';
+
+import { COINS_INITIAL_STATE } from './constants';
 import { checkCanAddMoney } from './utils/utils';
 
-export interface coins {
-  coin500: number;
-  coin100: number;
-  coin50: number;
-  coin10: number;
-}
-
-interface CoinManager {
-  getCoins(): coins;
-  getTotalAmount(): number;
-  addCoins(newCoins: coins): void;
-}
-
 export default class VendingMachineCoinManager implements CoinManager {
-  private coins: coins = {
-    coin500: 0,
-    coin100: 0,
-    coin50: 0,
-    coin10: 0,
-  };
+  private coins: coins = { ...COINS_INITIAL_STATE };
 
   getCoins() {
     return this.coins;
@@ -28,7 +13,7 @@ export default class VendingMachineCoinManager implements CoinManager {
   getTotalAmount() {
     return Object.entries(this.coins).reduce(
       (sum: number, [coin, count]: [string, number]) =>
-        sum + Number(coin.replace('coin', '')) * count,
+        sum + Number(coin.replace('COIN_', '')) * count,
       0
     );
   }
